rearanged conftest.py file and split it up
